requirements.txt
streamlit
Pillow
pylibdmtx